requirements.txt
requests
pandas
geopandas
shapely
python-dotenv
openpyxl
flask
gunicorn
